feat: clash邮件不再添加附件,改为使用订阅链接的方式.

+++ src/main/java/com/jiangzhiyan/clash/config/mail/sender/common/utils/MailUtil.java
@@ -50,25 +50,25 @@ public class MailUtil {
             return;
         }
 
-        List<String> emailContent = configType == ConfigType.CLASH
-                ? List.of("<strong>作者:</strong> " + commitAuthor,
-                "<strong>更新描述:</strong> " + commitMessage,
-                "<strong>文档:</strong> <a href='" + configType.getDocUrl() + "' target='_blank'>查看教程</a>",
-                "<strong>订阅链接:</strong> <a href='" + MessageFormat.format(CLASH_CONFIG_URL_TEMPLATE, branchName) + "' target='_blank'><i style='color: #228B22;'><u>打开然后复制链接url</u></i></a>",
-                "<i style='color: #999; font-size: smaller;'>此邮件由机器人自动发出，无需回复。</i>")
-                : List.of("<strong>作者:</strong> " + commitAuthor,
-                "<strong>更新描述:</strong> " + commitMessage,
-                "<strong>文档:</strong> <a href='" + configType.getDocUrl() + "' target='_blank'>查看教程</a>",
-                "<i style='color: #999; font-size: smaller;'>此邮件由机器人自动发出，无需回复。</i>");
+        // 邮件文字内容
+        MimeBodyPart textBodyPart = getEmailContent(branchName, configType, commitAuthor, commitMessage);
 
-        MimeBodyPart textBodyPart = new MimeBodyPart();
-        textBodyPart.setContent(String.join("<br/><br/>", emailContent), "text/html; charset=utf-8");
-
-        // 创建临时文件
-        String fileName = LocalDateTime.now().format(DATE_FORMATTER2) + "_config" + configType.getSuffix();
-        File tempFile = File.createTempFile(LocalDateTime.now().format(DATE_FORMATTER2) + "_config", configType.getSuffix());
-        try (FileWriter writer = new FileWriter(tempFile)) {
-            writer.write(fileContent);
+        // 文件内容,clash使用订阅链接,不使用配置文件
+        File tempFile = null;
+        MimeBodyPart attachmentBodyPart;
+        if (configType == ConfigType.CLASH) {
+            // 创建临时文件
+            String fileName = LocalDateTime.now().format(DATE_FORMATTER2) + "_config" + configType.getSuffix();
+            tempFile = File.createTempFile(LocalDateTime.now().format(DATE_FORMATTER2) + "_config", configType.getSuffix());
+            try (FileWriter writer = new FileWriter(tempFile)) {
+                writer.write(fileContent);
+            }
+            attachmentBodyPart = new MimeBodyPart();
+            DataSource source = new FileDataSource(tempFile);
+            attachmentBodyPart.setDataHandler(new DataHandler(source));
+            attachmentBodyPart.setFileName(fileName);
+        } else {
+            attachmentBodyPart = null;
         }
 
         Session session = Session.getInstance(PROPS, new Authenticator() {
@@ -89,14 +89,11 @@ public class MailUtil {
                     message.setRecipients(Message.RecipientType.TO, InternetAddress.parse(toEmail));
                     message.setSubject(LocalDateTime.now().format(DATE_FORMATTER).concat("_" + configType.getName() + "配置更新通知"));
 
-                    MimeBodyPart attachmentBodyPart = new MimeBodyPart();
-                    DataSource source = new FileDataSource(tempFile);
-                    attachmentBodyPart.setDataHandler(new DataHandler(source));
-                    attachmentBodyPart.setFileName(fileName);
-
                     Multipart multipart = new MimeMultipart();
                     multipart.addBodyPart(textBodyPart);
-                    multipart.addBodyPart(attachmentBodyPart);
+                    if (attachmentBodyPart != null) {
+                        multipart.addBodyPart(attachmentBodyPart);
+                    }
 
                     message.setContent(multipart);
 
@@ -120,11 +117,38 @@ public class MailUtil {
         } catch (InterruptedException | ExecutionException e) {
             throw new RuntimeException(e);
         } finally {
-            if (tempFile.exists()) {
+            if (tempFile != null && tempFile.exists()) {
                 if (!tempFile.delete()) {
                     System.err.println("Failed to delete temporary file: " + tempFile.getAbsolutePath());
                 }
             }
         }
+    }
+
+    /**
+     * 获取邮件正文内容
+     *
+     * @param branchName    分支名
+     * @param configType    配置类型
+     * @param commitAuthor  提交者
+     * @param commitMessage 提交信息
+     * @return 邮件正文内容
+     * @throws MessagingException 正文内容获取异常
+     */
+    private static MimeBodyPart getEmailContent(String branchName, ConfigType configType, String commitAuthor, String commitMessage) throws MessagingException {
+        List<String> emailContent = configType == ConfigType.CLASH
+                ? List.of("<strong>作者:</strong> " + commitAuthor,
+                "<strong>更新描述:</strong> " + commitMessage,
+                "<strong>文档:</strong> <a href='" + configType.getDocUrl() + "' target='_blank'>查看教程</a>",
+                "<strong>订阅链接:</strong> <a href='" + MessageFormat.format(CLASH_CONFIG_URL_TEMPLATE, branchName) + "' target='_blank'><i style='color: #228B22;'><u>打开然后复制链接url</u></i></a>",
+                "<i style='color: #999; font-size: smaller;'>此邮件由机器人自动发出，无需回复。</i>")
+                : List.of("<strong>作者:</strong> " + commitAuthor,
+                "<strong>更新描述:</strong> " + commitMessage,
+                "<strong>文档:</strong> <a href='" + configType.getDocUrl() + "' target='_blank'>查看教程</a>",
+                "<i style='color: #999; font-size: smaller;'>此邮件由机器人自动发出，无需回复。</i>");
+
+        MimeBodyPart textBodyPart = new MimeBodyPart();
+        textBodyPart.setContent(String.join("<br/><br/>", emailContent), "text/html; charset=utf-8");
+        return textBodyPart;
     }
 }
